[238] moved tests to one file
